:star: ADD: Feature Sigmoid Function(#7)
Menambahkan fungsi Sigmoid dan merubah alur pemanggilan activation
function
