+++ .gitignore
@@ -1,5 +1,5 @@
 # Compiled class file
-# *.class
+*.class
 
 # Log file
 *.log
